Move parseFacets to Facets component
