+++ frontend/src/components/Facets/index.jsx
@@ -8,13 +8,36 @@ import Alert from '../Feedback/Alert';
 import Divider from '../General/Divider';
 import Button from '../General/Button';
 import Spin from '../Feedback/Spin';
-import { isEmpty, humanize, parseFacets } from '../../utils/utils';
+
+import { isEmpty, humanize } from '../../utils/utils';
 import { fetchBaseFacets, fetchProjects } from '../../utils/api';
 
 const { Option } = Select;
 
 const styles = {
   facetCount: { float: 'right' },
+};
+
+/**
+ * Joins adjacent elements of the facets obj into a tuple using reduce().
+ * https://stackoverflow.com/questions/37270508/javascript-function-that-converts-array-to-array-of-2-tuples
+ * @param {Object.<string, Array.<Array<string, number>>} facets
+ */
+const parseFacets = (facets) => {
+  const res = facets;
+  const keys = Object.keys(facets);
+
+  keys.forEach((key) => {
+    res[key] = res[key].reduce((r, a, i) => {
+      if (i % 2) {
+        r[r.length - 1].push(a);
+      } else {
+        r.push([a]);
+      }
+      return r;
+    }, []);
+  });
+  return res;
 };
 
 function Facets({ project, onProjectChange, onSetFacets }) {
