+++ frontend/src/utils/utils.js
@@ -17,29 +17,3 @@ export function humanize(str) {
   }
   return frags.join(' ');
 }
-
-/**
- * Joins adjacent elements of the facets obj into a tuple using reduce().
- * setParsedFacets converts an array to an array of tuples.
- * For example, the array [var1, count1, var2, count2] turns into
- * [[var1, count1], [var2, count2]]
- *
- * Source: https://stackoverflow.com/questions/37270508/javascript-function-that-converts-array-to-array-of-2-tuples
- * @param {Object.<string, Array.<Array<string, number>>} facets
- */
-export function parseFacets(facets) {
-  const res = facets;
-  const keys = Object.keys(facets);
-
-  keys.forEach((key) => {
-    res[key] = res[key].reduce((r, a, i) => {
-      if (i % 2) {
-        r[r.length - 1].push(a);
-      } else {
-        r.push([a]);
-      }
-      return r;
-    }, []);
-  });
-  return res;
-}
